Refactoring supported language integration
